reorganize components according to team decision
