+++ src/components/GitHubAvatar/GitHubAvatar.tsx
