+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import RepoDropdown from '@/components/RepoDropdown';
+import RepoDropdown from '@/components/Header/RepoDropdown';
 
 export default function Header() {
   return (
