+++ src/components/Sidebar/AuthButton.tsx
