+++ src/components/Sidebar/SidebarToggle.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import HamburgerIcon from '@/components/icons/HamburgerIcon';
+import HamburgerIcon from '@/icons/HamburgerIcon';
 export default function SidebarToggle({
   sidebarCheckboxId,
 }: {
