+++ src/icons/FailedXIcon.tsx
