+++ src/icons/QualityIcon.tsx
