+++ src/icons/SignOutIcon.tsx
